--- a/builder/src/webpack.config.base.ts
+++ b/builder/src/webpack.config.base.ts
@@ -4,7 +4,8 @@ import * as webpack from 'webpack';
 import miniSVGDataURI from 'mini-svg-data-uri';
 
 const rules = [
-  { test: /\.css$/, use: ['style-loader', 'css-loader'] },
+  { test: /\.raw\.css$/, type: 'asset/source' },
+  { test: /(?<!\.raw)\.css$/, use: ['style-loader', 'css-loader'] },
   { test: /\.txt$/, type: 'asset/source' },
   { test: /\.md$/, type: 'asset/source' },
   { test: /\.(jpg|png|gif)$/, type: 'asset/resource' },