--- a/packages/apputils-extension/src/typings.d.ts
+++ b/packages/apputils-extension/src/typings.d.ts
@@ -1,0 +1,4 @@
+declare module '*.raw.css' {
+  const script: string;
+  export default script;
+}