--- a/packages/apputils-extension/src/typings.d.ts
+++ b/packages/apputils-extension/src/typings.d.ts
@@ -1,3 +1,8 @@
+/*
+ * Copyright (c) Jupyter Development Team.
+ * Distributed under the terms of the Modified BSD License.
+ */
+
 declare module '*.raw.css' {
   const script: string;
   export default script;