## References

#14167, #9757

Reduces switch time to a very large notebook in Chromium browsers by ~0.5s (tested on [`gh-9757-reproducer.ipynb`](https://gist.github.com/krassowski/eb2b312576e8597f935817c4c4d81fa5) which is representative of a notebook with outputs rather than of a long notebook with few visualisations).

## Code changes

By loading  `::-webkit-scrollbar` rules manually only when needed the benefit is twofold:
- when styling scrollbars is off we exclude problematic selectors
- when styling is on the time needed to match the selector is greatly reduced from `O(n * d)` to `O(n)` (equivalent to single universal selector), where:
   - `n` = number of DOM nodes
   - `d` = expected depth of the DOM tree (we can naively assume `log(n)`).

## User-facing changes

| Recalculate Styles + Layout time | Before | After | Faster by |
| --- | --- | --- | --- |
| Scrollbar styling on | 2.46s + 1.35s | 1.91s + 0.97s | 0.93s (24%) |
| Scrollbar styling off | 2.35s + 1.27s | 2.00s + 1.11s | 0.51s (14%) |

All results recorded with dark theme, using Edge Version 113.0.1754.0 (Official build) dev (64-bit) with advanced rendering instrumentation enabled.

There is no performance difference on Firefox.

<details>

### Before proposed, styling off

| Elapsed (µs) | Fast Reject Count | Match Attempts | Match Count | Selector |
-- | -- | -- | -- | --
134014 | 0 | 150287 | 0 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar
113392 | 0 | 150287 | 0 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar-thumb
112935 | 0 | 150287 | 0 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar-corner
20610 | 287 | 150287 | 3 | .jp-RenderedHTMLCommon > :last-child
19033 | 0 | 150287 | 150284 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar-track:horizontal
15133 | 0 | 150287 | 150284 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar-track:vertical
9260 | 0 | 50017 | 50017 | svg:not(:root)
7818 | 150287 | 150287 | 0 | .jp-MainAreaWidget-ContainStrict .jp-Notebook *
7755 | 150287 | 150287 | 0 | mjx-row > *
7662 | 150283 | 150287 | 4 | .ͼ1 .cm-layer > *
7624 | 150245 | 150287 | 42 | .ͼ4 .cm-line ::selection
7507 | 150233 | 150287 | 54

### Before proposed, styling on

| Elapsed (µs) | Fast Reject Count | Match Attempts | Match Count | Selector |
-- | -- | -- | -- | --
140145 | 0 | 150287 | 150284 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar
114441 | 0 | 150287 | 150284 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar-thumb
112257 | 0 | 150287 | 150284 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar-corner
21266 | 287 | 150287 | 3 | .jp-RenderedHTMLCommon > :last-child
15074 | 0 | 150287 | 150284 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar-track:horizontal
14780 | 0 | 150287 | 150284 | [data-jp-theme-scrollbars="true"] ::-webkit-scrollbar-track:vertical
9879 | 0 | 50017 | 50017 | svg:not(:root)
8599 | 150287 | 150287 | 0 | .jp-LSPExtension-FormGroup-content > *
8148 | 150287 | 150287 | 0 | .jp-RenderedJSON :not(mark)
8105 | 150287 | 150287 | 0 | .jp-MainAreaWidget-ContainStrict .jp-Notebook *
7909 | 0 | 50017 | 50017 | svg:not(:root)
7756 | 150287 | 150287 | 0 | mjx-row > *

### After proposed, styling off

| Elapsed (µs) | Fast Reject Count | Match Attempts | Match Count | Selector |
-- | -- | -- | -- | --
21856 | 287 | 150287 | 3 | .jp-RenderedHTMLCommon > :last-child
9493 | 0 | 50017 | 50017 | svg:not(:root)
8263 | 150287 | 150287 | 0 | .jp-MainAreaWidget-ContainStrict .jp-Notebook *
8087 | 150287 | 150287 | 0 | .jp-LSPExtension-FormGroup-content > *
8044 | 150287 | 150287 | 0 | mjx-row > *
8029 | 150233 | 150287 | 54 | .ͼo .cm-content ::selection
7935 | 150287 | 150287 | 0 | .jp-RenderedJSON :not(mark)
7848 | 150287 | 150287 | 0 | .jp-RankedPanel > :not(:first-child)
7786 | 150245 | 150287 | 42 | .ͼ4 .cm-line ::selection
7773 | 150283 | 150287 | 4 | .ͼ1 .cm-layer > *
7654 | 0 | 50017 | 50017 | svg:not(:root)
5886 | 0 | 50058 | 41 | :not(svg)
5753 | 17 | 50017 | 50000 | .jp-RenderedHTMLCommon svg
5385 | 0 | 50143 | 50142 | div
5228 | 0 | 50017 | 0 | :not(foreignObject) > svg
3023 | 50017 | 50017 | 0 | .fllbajr svg

### After proposed, styling on

| Elapsed (µs) | Fast Reject Count | Match Attempts | Match Count | Selector |
-- | -- | -- | -- | --
21359 | 293 | 150293 | 3 | .jp-RenderedHTMLCommon > :last-child
20157 | 0 | 150293 | 150290 | ::-webkit-scrollbar
15898 | 0 | 150293 | 150290 | ::-webkit-scrollbar-thumb
14612 | 0 | 150293 | 150290 | ::-webkit-scrollbar-track:horizontal
14592 | 0 | 150293 | 150290 | ::-webkit-scrollbar-track:vertical
14193 | 0 | 150293 | 150290 | ::-webkit-scrollbar-corner
9277 | 0 | 50018 | 50018 | svg:not(:root)
7933 | 150251 | 150293 | 42 | .ͼ4 .cm-line ::selection
7876 | 150293 | 150293 | 0 | .jp-MainAreaWidget-ContainStrict .jp-Notebook *
7661 | 150293 | 150293 | 0 | .jp-RenderedJSON :not(mark)
7646 | 150239 | 150293 | 54 | .ͼo .cm-content ::selection
7574 | 150289 | 150293 | 4 | .ͼ1 .cm-layer > *
7569 | 0 | 50018 | 50018 | svg:not(:root)
7552 | 150293 | 150293 | 0 | mjx-row > *
7538 | 150293 | 150293 | 0 | .jp-LSPExtension-FormGroup-content > *
7508 | 150293 | 150293 | 0 | .jp-RankedPanel > :not(:first-child)

</details>


## Backwards-incompatible changes

None. Since the data attribute is preserved this could potentially be backported (but webpack config would need manual bakckporting).