@krassowski 2023-03-27
Marking this as draft as I would like to wait until Chrome 112 final is out to test it again before we decide on merging this (but feel free to review otherwise).