@krassowski 2023-03-29
Undrafting after testing with Chrome Version 113.0.5668.0 (Official Build) dev (64-bit) which is newer than version 112.0.5615.39 which already contains a fix for [Chromium bug 1425169](https://bugs.chromium.org/p/chromium/issues/detail?id=1425169) - and indeed the style calculation step is ~1 second faster, but this PR still retains merit of >10% improvement:

| Recalculate Styles + Layout time | Before | After | Faster by |
| --- | --- | --- | --- |
| Scrollbar styling on | 1.26s + 0.91s | 0.93s + 0.99s | 0.25s (12%) |
| Scrollbar styling off | 1.34s + 0.84s |  0.94s + 0.78s | 0.46s (21%) |